docker: Update build script - fix build with new patches and Mattermost v10.8.0
Not sure if it's the patch rework or the some change to the source in
v10.8.0, but now when building the "external_imports.go" file is
always included, including the closed-source components that we don't
have and causing build errors.

To fix, in the only way I could find, remove the file before starting
the build.